+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - **Fix:** "Show user's signature once per conversation" option is not respected (#18)
 - **Fix:** Current page not being calculated correctly when adding last message of conversation (#19)
 - **Fix:** Signature once status is not respected after adding message using quick reply (#20)
+- **Fix:** Inline editing message does not set show signature status correctly (#21)
 
 ## 1.2.0 (`1020070`)